pyproject.toml
[project]
name = "co"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "websocket-client>=1.8.0",
]
